pyproject.toml
[build-system]
requires = ["hatchling < 1.22"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dewret"]

[tool.portray]
modules = ["dewret"]
extra_markdown_extensions = [
  "pymdownx.inlinehilite",
  "pymdownx.snippets",
  "pymdownx.highlight",
]

[tool.portray.pdocs]
exclude_source = true

[tool.portray.mkdocs]
markdown_extensions = [
  { "pymdownx.superfences" = {} }
]

[tool.ruff.lint]
select = ["D", "F", "B"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[project]
name = "dewret"
description = "DEclarative Workflow REndering Tool"
homepage = "https://github.com/flaxandteal/dewret"
repository = "https://github.com/flaxandteal/dewret"
documentation = "https://flaxandteal.github.io/dewret"
readme = "README.md"
version = "0.10.0"
authors = [
  {name = "Phil Weir", email = "phil.weir@flaxandteal.co.uk"},
  {name = "Chris Nixon", email = "chris.nixon@sigma-sc.co.uk"},
  {name = "Ellery Ames", email = "ellery.ames@flaxandteal.co.uk"},
  {name = "Sharon Ingram", email = "sharon.ingram@flaxandteal.co.uk"},
]
license = "Apache-2.0"
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.11"
dependencies = ["dask >= 2022", "pyyaml", "attrs", "click", "setuptools >= 65.5.1", "numpy", "sympy"]

[tool.hatch.metadata]
allow-direct-references = true

[project.optional-dependencies]
test = ["pytest", "coverage", "ruff", "mypy"]
docs = [
  "portray @ git+https://github.com/flaxandteal/portray.git#v1.8.0-fat"
]

[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64", "osx-arm64", "osx-64", "win-64"]
